requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
comm==0.1.3
crispy-bootstrap5==0.7
cryptography==39.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.52.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
executing==1.2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg==3.1.8
psycopg-binary==3.1.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pytest==7.3.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pyzmq==25.1.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
whitenoise==6.4.0
